requirements.txt
asgiref==3.7.2
Brotli==1.1.0
dj-database-url==0.5.0
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
whitenoise==6.6.0
